SBOM_Columns_Structures/requirements.txt
blinker===1.7.0
click===8.1.7
colorama===0.4.6
contourpy===1.2.0
cycler===0.12.1
easyocr===1.7.1
et-xmlfile===1.1.0
filelock===3.13.1
Flask==3.0.2
Flask-WTF===1.2.1
fonttools===4.49.0
fsspec===2024.2.0
imageio===2.34.0
imutils===0.5.4
itsdangerous===2.1.2
Jinja2===3.1.3
kiwisolver===1.4.5
lazy_loader===0.3
MarkupSafe===2.1.5
matplotlib===3.8.3
mpmath===1.3.0
networkx===3.2.1
ninja===1.11.1.1
numpy===1.26.4
opencv-python===4.9.0.80
opencv-python-headless===4.9.0.80
openpyxl===3.1.2
packaging===24.0
pdf2image===1.17.0
pillow===10.2.0
pyclipper===1.3.0.post5
pyparsing===3.1.2
python-bidi===0.4.2
python-dateutil===2.9.0.post0
PyYAML===6.0.1
scikit-image===0.22.0
scipy===1.12.0
setuptools===69.1.1
shapely===2.0.3
six===1.16.0
sympy===1.12
tifffile===2024.2.12
torch===2.2.1
torchaudio===2.2.1
torchvision===0.17.1
typing_extensions===4.10.0
Werkzeug===3.0.1
wheel===0.42.0
WTForms===3.1.2
